@@ -12,27 +12,27 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^10.0.14",
-        "@angular/cdk": "^10.2.3",
-        "@angular/common": "^10.0.14",
-        "@angular/compiler": "^10.0.14",
-        "@angular/core": "^10.0.14",
-        "@angular/forms": "^10.0.14",
-        "@angular/material": "^10.2.3",
-        "@angular/platform-browser": "^10.0.14",
-        "@angular/platform-browser-dynamic": "^10.0.14",
-        "@angular/router": "^10.0.14",
-        "karma-viewport": "^1.0.7",
-        "rxjs": "~6.5.5",
+        "@angular/animations": "^10.2.2",
+        "@angular/cdk": "^10.2.7",
+        "@angular/common": "^10.2.2",
+        "@angular/compiler": "^10.2.2",
+        "@angular/core": "^10.2.2",
+        "@angular/forms": "^10.2.2",
+        "@angular/material": "^10.2.7",
+        "@angular/platform-browser": "^10.2.2",
+        "@angular/platform-browser-dynamic": "^10.2.2",
+        "@angular/router": "^10.2.2",
+        "multer": "^1.4.2",
+        "rxjs": "^6.5.5",
         "tslib": "^2.0.0",
         "zone.js": "~0.10.2"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^0.1000.8",
+        "@angular-devkit/build-angular": "^0.1002.0",
         "@angular/cli": "^10.0.8",
-        "@angular/compiler-cli": "^10.0.14",
-        "@angular/language-service": "^10.0.14",
-        "@types/jasmine": "^3.5.14",
+        "@angular/compiler-cli": "^10.2.2",
+        "@angular/language-service": "^10.2.2",
+        "@types/jasmine": "^3.6.1",
         "@types/jasminewd2": "~2.0.3",
         "@types/node": "^12.12.62",
         "codelyzer": "^6.0.0",
@@ -44,6 +44,8 @@
         "karma-firefox-launcher": "^1.3.0",
         "karma-jasmine": "~3.3.0",
         "karma-jasmine-html-reporter": "^1.5.4",
+        "karma-viewport": "^1.0.7",
+        "ng-mocks": "^10.5.2",
         "prettier": "^2.1.2",
         "protractor": "~7.0.0",
         "ts-node": "~8.3.0",
