@@ -4,7 +4,6 @@ import * as cors from 'cors';
 import * as express from 'express';
 import { inject, injectable } from 'inversify';
 import * as logger from 'morgan';
-import { DateController } from './controllers/date.controller';
 import { IndexController } from './controllers/index.controller';
 import { TYPES } from './types';
 
@@ -13,10 +12,7 @@ export class Application {
     private readonly internalError: number = 500;
     app: express.Application;
 
-    constructor(
-        @inject(TYPES.IndexController) private indexController: IndexController,
-        @inject(TYPES.DateController) private dateController: DateController,
-    ) {
+    constructor(@inject(TYPES.IndexController) private indexController: IndexController) {
         this.app = express();
 
         this.config();
@@ -36,7 +32,6 @@ export class Application {
     bindRoutes(): void {
         // Notre application utilise le routeur de notre API `Index`
         this.app.use('/api/index', this.indexController.router);
-        this.app.use('/api/date', this.dateController.router);
         this.errorHandling();
     }
 
