@@ -1,7 +1,9 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { Tool } from '@app/classes/tool';
+import { ColorSelectionService } from '@app/services/color/color-selection-service';
 import { DrawingService } from '@app/services/drawing/drawing.service';
 import { ToolsService } from '@app/services/tools/tools-service';
+import { UndoRedoService } from '@app/services/tools/undo-redo-service';
 import { DrawingComponent } from './drawing.component';
 
 import SpyObj = jasmine.SpyObj;
@@ -15,19 +17,21 @@ describe('DrawingComponent', () => {
     let drawingStub: DrawingService;
     let toolsServiceSpy: SpyObj<ToolsService>;
 
-    beforeEach(async(() => {
-        toolStub = new ToolStub({} as DrawingService);
-        drawingStub = new DrawingService();
-        toolsServiceSpy = jasmine.createSpyObj('ToolsService', ['']);
-        toolsServiceSpy.currentDrawingTool = toolStub;
-        TestBed.configureTestingModule({
-            declarations: [DrawingComponent],
-            providers: [
-                { provide: ToolsService, useValue: toolsServiceSpy },
-                { provide: DrawingService, useValue: drawingStub },
-            ],
-        }).compileComponents();
-    }));
+    beforeEach(
+        waitForAsync(() => {
+            toolStub = new ToolStub({} as DrawingService, {} as UndoRedoService, {} as ColorSelectionService);
+            drawingStub = new DrawingService();
+            toolsServiceSpy = jasmine.createSpyObj('ToolsService', ['']);
+            toolsServiceSpy.currentDrawingTool = toolStub;
+            TestBed.configureTestingModule({
+                declarations: [DrawingComponent],
+                providers: [
+                    { provide: ToolsService, useValue: toolsServiceSpy },
+                    { provide: DrawingService, useValue: drawingStub },
+                ],
+            }).compileComponents();
+        }),
+    );
 
     beforeEach(() => {
         fixture = TestBed.createComponent(DrawingComponent);
