@@ -1,10 +1,13 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { MatDialogModule } from '@angular/material/dialog';
+import { MatSidenavModule } from '@angular/material/sidenav';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { AttributePanelComponent } from '@app/components/attribute-panel/attribute-panel.component';
 import { DrawingComponent } from '@app/components/drawing/drawing.component';
 import { SidebarComponent } from '@app/components/sidebar/sidebar.component';
 import { MIN_CANVAS_SIZE } from '@app/shared/constant';
+import { MockComponents } from 'ng-mocks';
 import { EditorComponent } from './editor.component';
 
 // tslint:disable:no-any
@@ -17,12 +20,14 @@ describe('EditorComponent', () => {
     let mouseEvent: MouseEvent;
     let setAnchorSpy: jasmine.Spy<any>;
 
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            imports: [MatDialogModule, BrowserModule, BrowserAnimationsModule],
-            declarations: [EditorComponent, DrawingComponent, SidebarComponent],
-        }).compileComponents();
-    }));
+    beforeEach(
+        waitForAsync(() => {
+            TestBed.configureTestingModule({
+                imports: [MatDialogModule, BrowserModule, BrowserAnimationsModule, MatSidenavModule],
+                declarations: [EditorComponent, MockComponents(SidebarComponent, AttributePanelComponent, DrawingComponent)],
+            }).compileComponents();
+        }),
+    );
 
     beforeEach(() => {
         fixture = TestBed.createComponent(EditorComponent);
