@@ -1,15 +1,20 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { GuideComponent } from './guide.component';
 
 describe('GuideComponent', () => {
     let component: GuideComponent;
     let fixture: ComponentFixture<GuideComponent>;
 
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            declarations: [GuideComponent],
-        }).compileComponents();
-    }));
+    beforeEach(
+        waitForAsync(() => {
+            TestBed.configureTestingModule({
+                declarations: [GuideComponent],
+            }).compileComponents();
+            fixture = TestBed.createComponent(GuideComponent);
+            component = fixture.componentInstance;
+            fixture.detectChanges();
+        }),
+    );
 
     beforeEach(() => {
         fixture = TestBed.createComponent(GuideComponent);
@@ -19,12 +24,5 @@ describe('GuideComponent', () => {
 
     it('should create', () => {
         expect(component).toBeTruthy();
-    });
-
-    it('should change tabContent and tabLink', () => {
-        const tab = 'crayon';
-        component.openTab(tab);
-        expect(component.tabcontent).toBeDefined();
-        expect(component.tablinks).toBeDefined();
     });
 });
