@@ -1,11 +1,16 @@
 import { Component, HostListener } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
+import { CarouselComponent } from '@app/components/carousel/carousel.component';
+import { CreateNewDrawingComponent } from '@app/components/create-new-drawing/create-new-drawing.component';
 import { GuideComponent } from '@app/components/guide/guide.component';
+import { ExportPopupComponent } from '@app/components/popup/export-popup/export-popup.component';
+import { SavePopupComponent } from '@app/components/popup/save-popup/save-popup.component';
+import { SidebarTool } from '@app/components/sidebar/sidebar-tool/sidebar-tool';
 import { DrawingService } from '@app/services/drawing/drawing.service';
 import { ToolsService } from '@app/services/tools/tools-service';
+import { UndoRedoService } from '@app/services/tools/undo-redo-service';
 import { drawingToolId, sidebarToolID } from '@app/shared/enum';
-import { SidebarToolComponent } from './sidebar-tool/sidebar-tool.component';
-// tslint:disable-next-line: no-relative-imports
+
 declare type callback = () => void;
 
 @Component({
@@ -14,65 +19,128 @@ declare type callback = () => void;
     styleUrls: ['./sidebar.component.scss'],
 })
 export class SidebarComponent {
-    sideBarToolsTop: SidebarToolComponent[];
-    sideBarToolsBottom: SidebarToolComponent[];
+    sideBarToolsTop: SidebarTool[];
+    sideBarToolsBottom: SidebarTool[];
+    sideBarUndoRedoButtons: SidebarTool[];
+    sideBarToolsTopMap: Map<sidebarToolID, SidebarTool> = new Map<sidebarToolID, SidebarTool>();
+    sideBarToolsBottomMap: Map<sidebarToolID, SidebarTool> = new Map<sidebarToolID, SidebarTool>();
 
     showDrawingTools: boolean;
+    isDialogOpen: boolean;
 
-    constructor(private toolsService: ToolsService, private dialog: MatDialog, private drawingService: DrawingService) {
+    constructor(
+        private toolsService: ToolsService,
+        private dialog: MatDialog,
+        private drawingService: DrawingService,
+        public undoRedo: UndoRedoService,
+    ) {
         this.sideBarToolsTop = [
-            { id: sidebarToolID.move, name: 'Select & Move' },
-            { id: sidebarToolID.selection, name: 'Selection' },
-            { id: sidebarToolID.tracing, name: 'Tracing' },
-            { id: sidebarToolID.shapes, name: 'Shapes' },
-            { id: sidebarToolID.line, name: 'Line' },
-            { id: sidebarToolID.text, name: 'Text' },
-            { id: sidebarToolID.paintBucket, name: 'Bucket' },
-            { id: sidebarToolID.stamp, name: 'Stamp' },
-            { id: sidebarToolID.pipette, name: 'Pipette' },
-            { id: sidebarToolID.eraser, name: 'Eraser' },
+            { id: sidebarToolID.selection, name: 'Selection', defaultDrawingToolid: drawingToolId.rectangleSelectionService },
+            { id: sidebarToolID.tracing, name: 'Traçage', defaultDrawingToolid: drawingToolId.pencilService },
+            { id: sidebarToolID.shapes, name: 'Figures', defaultDrawingToolid: drawingToolId.rectangleService },
+            { id: sidebarToolID.line, name: 'Ligne', defaultDrawingToolid: drawingToolId.lineService },
+            { id: sidebarToolID.text, name: 'Texte' },
+            { id: sidebarToolID.paintBucket, name: 'Sceau', defaultDrawingToolid: drawingToolId.bucketService },
+            { id: sidebarToolID.stamp, name: 'Étampe' },
+            { id: sidebarToolID.pipette, name: 'Pipette', defaultDrawingToolid: drawingToolId.pipetteService },
+            { id: sidebarToolID.eraser, name: 'Efface', defaultDrawingToolid: drawingToolId.eraserService },
         ];
+
+        this.sideBarUndoRedoButtons = [
+            { id: sidebarToolID.undo, name: 'annuler' },
+            { id: sidebarToolID.redo, name: 'refaire' },
+        ];
+
         this.sideBarToolsBottom = [
-            { id: sidebarToolID.createNew, name: 'New Drawing' },
-            { id: sidebarToolID.saveCurrent, name: 'Save Drawing' },
-            { id: sidebarToolID.openCarrousel, name: 'Open Carrousel' },
-            { id: sidebarToolID.exportCurrent, name: 'Export Drawing' },
-            { id: sidebarToolID.openGuide, name: 'Open Guide' },
+            { id: sidebarToolID.createNew, name: 'Nouveau dessin' },
+            { id: sidebarToolID.saveCurrent, name: 'Sauvegarder' },
+            { id: sidebarToolID.openCarrousel, name: 'Carroussel' },
+            { id: sidebarToolID.exportCurrent, name: 'Exporter' },
+            { id: sidebarToolID.openGuide, name: 'Ouvrir Guide' },
         ];
+        this.sideBarToolsTop.forEach((object) => {
+            this.sideBarToolsTopMap.set(object.id, object);
+        });
+        this.sideBarToolsBottom.forEach((object) => {
+            this.sideBarToolsBottomMap.set(object.id, object);
+        });
     }
 
-    onButtonPress(id: sidebarToolID): void {
+    openCloseSidenav(id: sidebarToolID): void {
+        if (this.toolsService._selectedSideBarToolID === id && this.toolsService.toolSidenavToggle.getValue() === true) {
+            this.toolsService.closeToolSidenav();
+        } else {
+            this.toolsService.openToolSidenav();
+        }
+    }
+
+    onButtonPressTop(object: SidebarTool | undefined): void {
+        if (object) {
+            this.openCloseSidenav(object.id);
+            this.toolsService._selectedSideBarToolID = object.id;
+            if (object.defaultDrawingToolid !== undefined) {
+                this.toolsService._currentDrawingTool = object.defaultDrawingToolid;
+            }
+        }
+    }
+
+    onButtonPressBottom(id: sidebarToolID): void {
         switch (id) {
-            case sidebarToolID.tracing: {
-                this.toolsService._selectedSideBarToolID = sidebarToolID.tracing;
-                this.toolsService._currentDrawingTool = drawingToolId.pencilService;
-                break;
-            }
-            case sidebarToolID.shapes: {
-                this.toolsService._selectedSideBarToolID = sidebarToolID.shapes;
-                this.toolsService._currentDrawingTool = drawingToolId.rectangleService;
-                break;
-            }
-            case sidebarToolID.line: {
-                this.toolsService._selectedSideBarToolID = sidebarToolID.line;
-                this.toolsService._currentDrawingTool = drawingToolId.lineService;
-                break;
-            }
-            case sidebarToolID.eraser: {
-                this.toolsService._selectedSideBarToolID = sidebarToolID.eraser;
-                this.toolsService._currentDrawingTool = drawingToolId.eraserService;
-                break;
-            }
             case sidebarToolID.createNew: {
-                this.drawingService.clearCanvas(this.drawingService.baseCtx);
-                this.drawingService.clearCanvas(this.drawingService.previewCtx);
+                this.undoRedo.clearPile();
+                this.isDialogOpen = true;
+                const dialogRef = this.dialog.open(CreateNewDrawingComponent);
+                dialogRef.afterClosed().subscribe(() => {
+                    this.isDialogOpen = false;
+                });
                 break;
             }
             case sidebarToolID.openGuide: {
+                this.isDialogOpen = true;
                 const dialogRef = this.dialog.open(GuideComponent);
+                dialogRef.afterClosed().subscribe(() => {
+                    this.isDialogOpen = false;
+                });
+                break;
+            }
+            case sidebarToolID.openCarrousel: {
+                this.isDialogOpen = true;
+                const dialogRef = this.dialog.open(CarouselComponent);
+                dialogRef.afterClosed().subscribe(() => {
+                    this.isDialogOpen = false;
+                });
+                break;
+            }
+            case sidebarToolID.saveCurrent: {
+                this.isDialogOpen = true;
+                const dialogRef = this.dialog.open(SavePopupComponent);
+                dialogRef.afterClosed().subscribe(() => {
+                    this.isDialogOpen = false;
+                });
+                break;
+            }
+            case sidebarToolID.exportCurrent: {
+                this.isDialogOpen = true;
+                const dialogRef = this.dialog.open(ExportPopupComponent);
+                dialogRef.afterClosed().subscribe(() => {
+                    this.isDialogOpen = false;
+                });
+                break;
+            }
+            case sidebarToolID.openCarrousel: {
+                this.isDialogOpen = true;
+                const dialogRef = this.dialog.open(CarouselComponent, { width: '90%', height: '70%' });
 
-                dialogRef.afterClosed().subscribe((result) => {
-                    console.log(`Dialog result: ${result}`);
+                dialogRef.afterClosed().subscribe(() => {
+                    this.isDialogOpen = false;
+                });
+                break;
+            }
+            case sidebarToolID.saveCurrent: {
+                const dialogRef = this.dialog.open(SavePopupComponent);
+                this.isDialogOpen = true;
+                dialogRef.afterClosed().subscribe(() => {
+                    this.isDialogOpen = false;
                 });
                 break;
             }
@@ -87,6 +155,10 @@ export class SidebarComponent {
         if (event.shiftKey) {
             keys += 'S-';
         }
+        if (event.shiftKey && event.ctrlKey) {
+            keys = 'C-';
+            keys += 'S-';
+        }
         keys += event.key.toLowerCase();
         return keys;
     }
@@ -99,30 +171,65 @@ export class SidebarComponent {
                 this.drawingService.clearCanvas(this.drawingService.baseCtx);
                 this.drawingService.clearCanvas(this.drawingService.previewCtx);
             },
+            'C-e': () => {
+                this.onButtonPressBottom(sidebarToolID.exportCurrent);
+            },
+            'C-g': () => {
+                this.onButtonPressBottom(sidebarToolID.openCarrousel);
+            },
+            'C-s': () => {
+                this.onButtonPressBottom(sidebarToolID.saveCurrent);
+            },
+            'C-z': () => {
+                this.undoRedo.undo();
+            },
+            'C-S-z': () => {
+                this.undoRedo.redo();
+            },
         };
         const func: callback | undefined = kbd[keys];
         if (func) {
-            event.preventDefault();
+            // event.preventDefault();
             func();
         }
     }
 
     @HostListener('window: keyup', ['$event'])
     onKeyUp(event: KeyboardEvent): void {
+        if (this.isDialogOpen) {
+            return;
+        }
         const kbd: { [id: string]: callback } = {
             c: () => {
-                this.onButtonPress(sidebarToolID.tracing);
-                this.toolsService._currentDrawingTool = drawingToolId.pencilService;
+                this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.tracing));
             },
-            e: () => this.onButtonPress(sidebarToolID.eraser),
-            l: () => this.onButtonPress(sidebarToolID.line),
+            w: () => {
+                this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.tracing));
+                this.toolsService._currentDrawingTool = drawingToolId.brushService;
+            },
+            b: () => this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.paintBucket)),
+            e: () => this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.eraser)),
+            l: () => this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.line)),
+            r: () => {
+                this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.selection));
+            },
+
+            s: () => {
+                this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.selection));
+                this.toolsService._currentDrawingTool = drawingToolId.ellipseSelectionService;
+            },
+            i: () => this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.pipette)),
+
             1: () => {
-                this.onButtonPress(sidebarToolID.shapes);
-                this.toolsService._currentDrawingTool = drawingToolId.rectangleService;
+                this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.shapes));
             },
             2: () => {
-                this.onButtonPress(sidebarToolID.shapes);
+                this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.shapes));
                 this.toolsService._currentDrawingTool = drawingToolId.ellipseService;
+            },
+            3: () => {
+                this.onButtonPressTop(this.sideBarToolsTopMap.get(sidebarToolID.shapes));
+                this.toolsService._currentDrawingTool = drawingToolId.polygonService;
             },
         };
         const keys: string = this.getComposedKey(event);
@@ -130,5 +237,13 @@ export class SidebarComponent {
             const func: callback = kbd[keys];
             func();
         }
+    }
+
+    get showUndo(): boolean {
+        return this.undoRedo.undoPile.length > 0;
+    }
+
+    get showRedo(): boolean {
+        return this.undoRedo.redoPile.length > 0;
     }
 }
