@@ -1,33 +1,37 @@
 import { Injectable } from '@angular/core';
+import { DrawingAction } from '@app/classes/drawing-action';
 import { Tool } from '@app/classes/tool';
+import { ToolOption } from '@app/classes/tool-option';
 import { Vec2 } from '@app/classes/vec2';
+import { ColorSelectionService } from '@app/services/color/color-selection-service';
 import { DrawingService } from '@app/services/drawing/drawing.service';
-import { MouseButton, Texture } from '@app/shared/enum';
+import { UndoRedoService } from '@app/services/tools/undo-redo-service';
+import { DEFAULT_OPTIONS, TEXTURES } from '@app/shared/constant';
+import { drawingToolId, MouseButton, Options } from '@app/shared/enum';
 
 @Injectable({
     providedIn: 'root',
 })
 export class BrushService extends Tool {
     private pathData: Vec2[];
-    // Todo: Attributs globaux
-    // private color: string;
-    // private opacity: number;
-    private thickness: number;
-    selectedTexture: Texture;
+    private availableTextures: string[];
 
-    constructor(drawingService: DrawingService) {
-        super(drawingService);
-        this.thickness = 1; // Remplacer par un observable
+    constructor(drawingService: DrawingService, undoRedoService: UndoRedoService, colorService: ColorSelectionService) {
+        super(drawingService, undoRedoService, colorService);
         this.clearPath();
-        this.selectedTexture = Texture.one;
+        this.setDefaultOptions();
+        this.availableTextures = TEXTURES;
     }
 
-    set _thickness(newThickness: number) {
-        this.thickness = newThickness;
-    }
-
-    get _thickness(): number {
-        return this.thickness;
+    setDefaultOptions(): void {
+        const toolOptions = new Map<Options, ToolOption>([
+            [Options.size, { value: DEFAULT_OPTIONS.size, displayName: 'Largeur' }],
+            [Options.texture, { value: DEFAULT_OPTIONS.texture, displayName: 'Texture' }],
+        ]);
+        this.options = {
+            primaryColor: this.primaryColor,
+            toolOptions,
+        };
     }
 
     onMouseDown(event: MouseEvent): void {
@@ -43,7 +47,9 @@ export class BrushService extends Tool {
         if (this.mouseDown) {
             const mousePosition = this.getPositionFromMouse(event);
             this.pathData.push(mousePosition);
-            this.drawBrush(this.drawingService.baseCtx, this.pathData);
+            const drawingAction = this.getDrawingAction();
+            this.undoRedoService.saveAction(drawingAction);
+            this.draw(this.drawingService.baseCtx, drawingAction);
         }
         this.mouseDown = false;
         this.clearPath();
@@ -53,27 +59,48 @@ export class BrushService extends Tool {
         if (this.mouseDown && event.buttons === MouseButton.Left) {
             const mousePosition = this.getPositionFromMouse(event);
             this.pathData.push(mousePosition);
-            // On dessine sur le canvas de prévisualisation et on l'efface à chaque déplacement de la souris
-            this.drawBrush(this.drawingService.previewCtx, this.pathData);
+            // We draw on the preview canvas and erase it each time the mouse is moved
+            this.draw(this.drawingService.previewCtx, this.getDrawingAction());
         }
         if (this.mouseDown && !(event.buttons === MouseButton.Left)) {
-            this.drawBrush(this.drawingService.baseCtx, this.pathData);
+            this.draw(this.drawingService.baseCtx, this.getDrawingAction());
         }
     }
 
-    private drawBrush(ctx: CanvasRenderingContext2D, path: Vec2[]): void {
-        this.drawingService.clearCanvas(this.drawingService.previewCtx);
-        ctx.beginPath();
-        ctx.lineCap = 'round';
-        ctx.lineWidth = this.thickness;
-        const image = new Image(1, 1);
-        image.src = this.selectedTexture;
-        const pattern = ctx.createPattern(image, 'repeat');
-        if (pattern !== null) ctx.strokeStyle = pattern;
-        for (const point of path) {
-            ctx.lineTo(point.x, point.y);
+    draw(ctx: CanvasRenderingContext2D, drawingAction: DrawingAction): void {
+        const size = drawingAction.options.toolOptions.get(Options.size);
+        const texture = drawingAction.options.toolOptions.get(Options.texture);
+        if (drawingAction.path && size && texture) {
+            this.drawingService.clearCanvas(this.drawingService.previewCtx);
+            ctx.beginPath();
+            ctx.lineCap = 'round';
+            ctx.lineWidth = size.value;
+            const image = new Image(1, 1);
+            image.src = this.availableTextures[texture.value];
+            const pattern = ctx.createPattern(image, 'repeat');
+            if (pattern !== null) ctx.strokeStyle = pattern;
+            for (const point of drawingAction.path) {
+                ctx.lineTo(point.x, point.y);
+            }
+            ctx.stroke();
+            ctx.globalCompositeOperation = 'color';
+            ctx.strokeStyle = drawingAction.options.primaryColor.getRgbString();
+            ctx.stroke();
+            ctx.globalCompositeOperation = 'source-over';
         }
-        ctx.stroke();
+    }
+
+    getDrawingAction(): DrawingAction {
+        const options = {
+            primaryColor: this.primaryColor,
+            toolOptions: this.copyToolOptionMap(this.options.toolOptions),
+        };
+
+        return {
+            id: drawingToolId.brushService,
+            path: this.pathData,
+            options,
+        };
     }
 
     private clearPath(): void {
