@@ -1,81 +1,89 @@
 import { Injectable } from '@angular/core';
+import { DrawingAction } from '@app/classes/drawing-action';
 import { Tool } from '@app/classes/tool';
+import { ToolOption } from '@app/classes/tool-option';
 import { Vec2 } from '@app/classes/vec2';
+import { ColorSelectionService } from '@app/services/color/color-selection-service';
 import { DrawingService } from '@app/services/drawing/drawing.service';
-import { BACKSPACE_KEY, BASE_SNAP_ANGLE, ESCAPE_KEY, MIDDLE_SNAP_ANGLE, SHIFT_KEY } from '@app/shared/constant';
+import { UndoRedoService } from '@app/services/tools/undo-redo-service';
+import { BACKSPACE_KEY, BASE_SNAP_ANGLE, DEFAULT_OPTIONS, ESCAPE_KEY, MIDDLE_SNAP_ANGLE, SHIFT_KEY } from '@app/shared/constant';
+import { drawingToolId, Options } from '@app/shared/enum';
 
 @Injectable({
     providedIn: 'root',
 })
 export class LineService extends Tool {
     private pathData: Vec2[];
+    private lineStarted: boolean;
     currentCoord: Vec2;
-    lineStarted: boolean;
-    initialCoord: Vec2;
-    private thickness: number;
 
-    constructor(drawingService: DrawingService) {
-        super(drawingService);
-        this.thickness = 0;
+    constructor(drawingService: DrawingService, undoRedoService: UndoRedoService, colorService: ColorSelectionService) {
+        super(drawingService, undoRedoService, colorService);
         this.clearPath();
+        this.setDefaultOptions();
         this.lineStarted = false;
     }
+    setDefaultOptions(): void {
+        const toolOptions = new Map<Options, ToolOption>([[Options.size, { value: DEFAULT_OPTIONS.size, displayName: 'Largeur' }]]);
 
-    set _thickness(newThickness: number) {
-        this.thickness = newThickness;
+        this.options = {
+            primaryColor: this.primaryColor,
+            toolOptions,
+        };
     }
 
-    get _thickness(): number {
-        return this.thickness;
+    private getLastClickedCoord(): Vec2 {
+        const length = this.pathData.length;
+        return length > 1 ? this.pathData[length - 2] : { x: 0, y: 0 };
     }
 
     onMouseClick(event: MouseEvent): void {
         if (this.lineStarted) {
-            this.pathData.push(this.currentCoord);
+            this.pathData.push(this.pathData[this.pathData.length - 1]);
         } else {
             this.lineStarted = true;
-            this.initialCoord = this.getPositionFromMouse(event);
-            this.currentCoord = this.initialCoord;
+            const initialCoord = this.getPositionFromMouse(event);
+            this.currentCoord = initialCoord;
+            this.pathData = [initialCoord, initialCoord];
         }
     }
 
     onMouseMove(event: MouseEvent): void {
         if (this.lineStarted) {
             this.currentCoord = this.getPositionFromMouse(event);
-            if (this.shiftDown) {
-                this.currentCoord = this.getSnappedCoord();
-            }
-            this.drawLine(this.drawingService.previewCtx);
+            this.updateLastCoord();
+            this.draw(this.drawingService.previewCtx, this.getDrawingAction());
         }
     }
 
     onMouseDoubleClick(event: MouseEvent): void {
         if (this.lineStarted) {
-            this.drawLine(this.drawingService.baseCtx);
             this.endLine();
+            const drawingAction = this.getDrawingAction();
+            this.undoRedoService.saveAction(drawingAction);
+            this.draw(this.drawingService.baseCtx, drawingAction);
+            this.lineStarted = false;
+            this.clearPath();
         }
     }
 
     endLine(): void {
         const closingMinDistance = 20;
-        const diff = this.getDiff(this.initialCoord, this.currentCoord);
+        const endPoint = this.getLastClickedCoord();
+        const initialPoint = this.pathData[0];
+        const diff = this.getDiff(initialPoint, endPoint);
         const closeShape = Math.abs(diff.x) <= closingMinDistance && Math.abs(diff.y) <= closingMinDistance;
         if (closeShape) {
-            const ctx = this.drawingService.baseCtx;
-            ctx.beginPath();
-            ctx.moveTo(this.currentCoord.x, this.currentCoord.y);
-            ctx.lineTo(this.initialCoord.x, this.initialCoord.y);
-            ctx.stroke();
+            this.pathData.splice(this.pathData.length - 1, 1, initialPoint);
         }
-        this.clearPath();
-        this.lineStarted = false;
     }
 
     onKeyUp(event: KeyboardEvent): void {
         if (event.key === SHIFT_KEY) {
             this.shiftDown = false;
             if (this.lineStarted) {
-                this.drawLine(this.drawingService.previewCtx);
+                this.updateLastCoord();
+                this.draw(this.drawingService.previewCtx, this.getDrawingAction());
             }
         }
     }
@@ -83,15 +91,18 @@ export class LineService extends Tool {
     onKeyDown(event: KeyboardEvent): void {
         switch (event.key) {
             case SHIFT_KEY:
-                this.shiftDown = true;
-                if (this.lineStarted) {
-                    this.drawLine(this.drawingService.previewCtx);
+                if (this.shiftDown !== true) {
+                    this.shiftDown = true;
+                    if (this.lineStarted) {
+                        this.updateLastCoord();
+                        this.draw(this.drawingService.previewCtx, this.getDrawingAction());
+                    }
                 }
                 break;
             case BACKSPACE_KEY:
-                this.pathData.pop();
                 if (this.lineStarted) {
-                    this.drawLine(this.drawingService.previewCtx);
+                    this.pathData.splice(this.pathData.length - 2, 1);
+                    this.draw(this.drawingService.previewCtx, this.getDrawingAction());
                 }
                 break;
             case ESCAPE_KEY:
@@ -102,58 +113,76 @@ export class LineService extends Tool {
         }
     }
 
+    updateLastCoord(): void {
+        const lastClickedCoord = this.getLastClickedCoord();
+        const adjustedCoord = this.getAdjustedCoord(lastClickedCoord, this.currentCoord);
+        this.pathData.splice(this.pathData.length - 1, 1, adjustedCoord);
+    }
+
     calculateAngle(startingPoint: Vec2, endPoint: Vec2): number {
         const diff = this.getDiff(startingPoint, endPoint);
         return Math.atan(Math.abs(diff.y) / Math.abs(diff.x));
     }
 
-    getSnappedCoord(): Vec2 {
-        const nPoints = this.pathData.length;
-        const startingPoint = nPoints > 0 ? this.pathData[nPoints - 1] : this.initialCoord;
-        const angle = this.calculateAngle(startingPoint, this.currentCoord);
+    getAdjustedCoord(initialPoint: Vec2, finalPoint: Vec2): Vec2 {
+        return this.shiftDown ? this.getSnappedCoord(initialPoint, finalPoint) : finalPoint;
+    }
+
+    getSnappedCoord(initialPoint: Vec2, finalPoint: Vec2): Vec2 {
+        const angle = this.calculateAngle(initialPoint, finalPoint);
 
         let point = { x: 0, y: 0 };
         if (angle < BASE_SNAP_ANGLE) {
-            point = { x: this.currentCoord.x, y: startingPoint.y };
+            point = { x: finalPoint.x, y: initialPoint.y };
         } else if (angle > MIDDLE_SNAP_ANGLE + BASE_SNAP_ANGLE) {
-            point = { x: startingPoint.x, y: this.currentCoord.y };
+            point = { x: initialPoint.x, y: finalPoint.y };
         } else {
-            point = this.getProjectedPoint(startingPoint, this.currentCoord, angle);
+            point = this.getProjectedPoint(initialPoint, finalPoint);
         }
         return point;
     }
 
-    getProjectedPoint(startingPoint: Vec2, endPoint: Vec2, angle: number): Vec2 {
-        const diff = this.getDiff(startingPoint, this.currentCoord);
-        // parameter angle is the angle between the line based on the two points and the positive x axis
-        // middle angle is between the line and the 45 degree line from the positive x axis
-        const middleAngle = MIDDLE_SNAP_ANGLE - angle;
-        const currentLineLenght = Math.sqrt(Math.pow(diff.x, 2) + Math.pow(diff.y, 2));
-        const projectedLineLenght = currentLineLenght * Math.cos(middleAngle);
+    getProjectedPoint(initialPoint: Vec2, finalPoint: Vec2): Vec2 {
+        const diff = this.getDiff(initialPoint, finalPoint);
+
         const projectedPoint = {
-            x: startingPoint.x + Math.cos(MIDDLE_SNAP_ANGLE) * projectedLineLenght * Math.sign(diff.x),
-            y: startingPoint.y + Math.sin(MIDDLE_SNAP_ANGLE) * projectedLineLenght * Math.sign(diff.y),
+            x: finalPoint.x,
+            y: initialPoint.y + Math.tan(MIDDLE_SNAP_ANGLE) * Math.abs(diff.x) * Math.sign(diff.y),
         };
         return projectedPoint;
     }
 
-    private drawLine(ctx: CanvasRenderingContext2D): void {
-        this.drawingService.clearCanvas(this.drawingService.previewCtx);
-        ctx.beginPath();
+    draw(ctx: CanvasRenderingContext2D, drawingAction: DrawingAction): void {
+        const size = drawingAction.options.toolOptions.get(Options.size);
+        if (drawingAction.path && size) {
+            this.drawingService.clearCanvas(this.drawingService.previewCtx);
+            ctx.beginPath();
+            ctx.lineWidth = size.value;
+            ctx.strokeStyle = drawingAction.options.primaryColor.getRgbString();
 
-        ctx.moveTo(this.initialCoord.x, this.initialCoord.y);
-        ctx.lineWidth = this.thickness;
-        for (const point of this.pathData) {
-            ctx.lineTo(point.x, point.y);
+            const startingPoint = drawingAction.path[0];
+            ctx.moveTo(startingPoint.x, startingPoint.y);
+            for (const point of drawingAction.path) {
+                ctx.lineTo(point.x, point.y);
+            }
+            ctx.stroke();
+            ctx.closePath();
         }
+    }
 
-        ctx.lineTo(this.currentCoord.x, this.currentCoord.y);
-        ctx.stroke();
+    getDrawingAction(): DrawingAction {
+        const options = {
+            primaryColor: this.primaryColor,
+            toolOptions: this.copyToolOptionMap(this.options.toolOptions),
+        };
+        return {
+            id: drawingToolId.lineService,
+            path: this.pathData,
+            options,
+        };
     }
 
     private clearPath(): void {
         this.pathData = [];
-        this.initialCoord = { x: 0, y: 0 };
-        this.currentCoord = { x: 0, y: 0 };
     }
 }
