@@ -1,63 +1,70 @@
 import { Injectable } from '@angular/core';
+import { BoundingBox } from '@app/classes/bounding-box';
+import { DrawingAction } from '@app/classes/drawing-action';
+import { SelectionBox } from '@app/classes/selection-box';
 import { Tool } from '@app/classes/tool';
+import { ToolOption } from '@app/classes/tool-option';
 import { Vec2 } from '@app/classes/vec2';
+import { ColorSelectionService } from '@app/services/color/color-selection-service';
 import { DrawingService } from '@app/services/drawing/drawing.service';
-import { SHIFT_KEY } from '@app/shared/constant';
-import { MouseButton, TraceTypes } from '@app/shared/enum';
+import { UndoRedoService } from '@app/services/tools/undo-redo-service';
+import { DEFAULT_OPTIONS, SHIFT_KEY } from '@app/shared/constant';
+import { drawingToolId, MouseButton, Options } from '@app/shared/enum';
 
 @Injectable({
     providedIn: 'root',
 })
 export class RectangleService extends Tool {
     initialCoord: Vec2;
-    private thickness: number;
-    private traceType: number;
+    selectionBox: SelectionBox;
 
-    constructor(drawingService: DrawingService) {
-        super(drawingService);
-        this.thickness = 0;
-        this.traceType = 0;
+    constructor(drawingService: DrawingService, undoRedoService: UndoRedoService, colorService: ColorSelectionService) {
+        super(drawingService, undoRedoService, colorService);
+        this.setDefaultOptions();
+        this.selectionBox = new SelectionBox();
     }
 
-    set _thickness(newThickness: number) {
-        this.thickness = newThickness;
-    }
-
-    get _thickness(): number {
-        return this.thickness;
-    }
-
-    set _traceType(newType: number) {
-        this.traceType = newType;
-    }
-
-    get _traceType(): number {
-        return this.traceType;
+    setDefaultOptions(): void {
+        const toolOptions = new Map<Options, ToolOption>([
+            [Options.size, { value: DEFAULT_OPTIONS.size, displayName: 'Largeur' }],
+            [Options.traceType, { value: DEFAULT_OPTIONS.traceType, displayName: 'Type' }],
+        ]);
+        this.options = {
+            primaryColor: this.primaryColor,
+            secondaryColor: this.secondaryColor,
+            toolOptions,
+        };
     }
 
     onMouseDown(event: MouseEvent): void {
         this.mouseDown = event.buttons === MouseButton.Left;
         if (this.mouseDown) {
             const currentCoord = this.getPositionFromMouse(event);
-            this.initialCoord = currentCoord;
-            this.mouseDownCoord = currentCoord;
+            this.selectionBox.setAnchor(currentCoord);
         }
     }
 
     onMouseUp(event: MouseEvent): void {
         if (this.mouseDown) {
-            this.drawRectangle(this.drawingService.baseCtx);
+            const currentCoord = this.getPositionFromMouse(event);
+            this.selectionBox.updateOpposingCorner(currentCoord);
+            const action = this.getDrawingAction();
+            this.undoRedoService.saveAction(action);
+            this.draw(this.drawingService.baseCtx, action);
         }
         this.mouseDown = false;
     }
 
     onMouseMove(event: MouseEvent): void {
         if (this.mouseDown && event.buttons === MouseButton.Left) {
-            this.mouseDownCoord = this.getPositionFromMouse(event);
-            this.drawRectangle(this.drawingService.previewCtx);
+            const currentCoord = this.getPositionFromMouse(event);
+            this.selectionBox.updateOpposingCorner(currentCoord);
+            this.draw(this.drawingService.previewCtx, this.getDrawingAction());
         }
         if (this.mouseDown && !(event.buttons === MouseButton.Left)) {
-            this.drawRectangle(this.drawingService.baseCtx);
+            const action = this.getDrawingAction();
+            this.undoRedoService.saveAction(action);
+            this.draw(this.drawingService.baseCtx, action);
             this.mouseDown = false;
         }
     }
@@ -66,7 +73,7 @@ export class RectangleService extends Tool {
         if (event.key === SHIFT_KEY) {
             this.shiftDown = false;
             if (this.mouseDown) {
-                this.drawRectangle(this.drawingService.previewCtx);
+                this.draw(this.drawingService.previewCtx, this.getDrawingAction());
             }
         }
     }
@@ -75,41 +82,37 @@ export class RectangleService extends Tool {
         if (event.key === SHIFT_KEY) {
             this.shiftDown = true;
             if (this.mouseDown) {
-                this.drawRectangle(this.drawingService.previewCtx);
+                this.draw(this.drawingService.previewCtx, this.getDrawingAction());
             }
         }
     }
 
-    private drawRectangle(ctx: CanvasRenderingContext2D): void {
-        this.drawingService.clearCanvas(this.drawingService.previewCtx);
-        ctx.beginPath();
-        // tslint:disable-next-line: prefer-const
-
-        const width = this.mouseDownCoord.x - this.initialCoord.x;
-        const height = this.mouseDownCoord.y - this.initialCoord.y;
-
-        ctx.lineWidth = this.thickness;
-
-        if (this.shiftDown) {
-            const squareSize = Math.min(Math.abs(width), Math.abs(height));
-            ctx.rect(this.initialCoord.x, this.initialCoord.y, squareSize * Math.sign(width), squareSize * Math.sign(height));
-        } else {
-            ctx.rect(this.initialCoord.x, this.initialCoord.y, width, height);
+    draw(ctx: CanvasRenderingContext2D, drawingAction: DrawingAction): void {
+        const options = drawingAction.options;
+        const size = options.toolOptions.get(Options.size);
+        const traceType = options.toolOptions.get(Options.traceType);
+        if (drawingAction.box && size && traceType && options.secondaryColor) {
+            this.drawingService.clearCanvas(this.drawingService.previewCtx);
+            ctx.beginPath();
+            ctx.lineWidth = size.value;
+            const box = drawingAction.box;
+            ctx.rect(box.position.x, box.position.y, box.width, box.height);
+            this.fill(ctx, traceType.value, options.primaryColor, options.secondaryColor);
         }
+    }
 
-        switch (this.traceType) {
-            case TraceTypes.fill:
-                ctx.fill();
-                break;
-            case TraceTypes.stroke:
-                ctx.stroke();
-                break;
-            case TraceTypes.fillAndStroke:
-                ctx.fill();
-                ctx.stroke();
-                break;
-            default:
-                ctx.fill();
-        }
+    getDrawingAction(): DrawingAction {
+        const options = {
+            primaryColor: this.primaryColor,
+            secondaryColor: this.secondaryColor,
+            toolOptions: this.copyToolOptionMap(this.options.toolOptions),
+        };
+        const box = new BoundingBox();
+        box.updateFromSelectionBox(this.selectionBox, this.shiftDown);
+        return {
+            id: drawingToolId.rectangleService,
+            box,
+            options,
+        };
     }
 }
