@@ -1,5 +1,7 @@
 import { TestBed } from '@angular/core/testing';
+import { ColorSelectionService } from '@app/services/color/color-selection-service';
 import { DrawingService } from '@app/services/drawing/drawing.service';
+import { UndoRedoService } from '@app/services/tools/undo-redo-service';
 import { drawingToolId } from '@app/shared/enum';
 import { PencilService } from './pencil-service';
 import { ToolsService } from './tools-service';
@@ -9,7 +11,7 @@ describe('ToolsService', () => {
     let pencilStub: PencilService;
 
     beforeEach(() => {
-        pencilStub = new PencilService({} as DrawingService);
+        pencilStub = new PencilService({} as DrawingService, {} as UndoRedoService, {} as ColorSelectionService);
         TestBed.configureTestingModule({
             providers: [{ provide: PencilService, useValue: pencilStub }],
         });
