@@ -1,12 +1,19 @@
 import { Injectable } from '@angular/core';
 import { Tool } from '@app/classes/tool';
+import { ToolOption } from '@app/classes/tool-option';
 import { BrushService } from '@app/services/tools/brush.service';
+import { BucketService } from '@app/services/tools/bucket-service';
 import { EllipseService } from '@app/services/tools/ellipse-service';
 import { EraserService } from '@app/services/tools/eraser-service';
 import { LineService } from '@app/services/tools/line-service';
 import { PencilService } from '@app/services/tools/pencil-service';
+import { PipetteService } from '@app/services/tools/pipette-service';
+import { PolygonService } from '@app/services/tools/polygon-service';
+import { RectangleSelectorService } from '@app/services/tools/rectangle-selector-service';
 import { RectangleService } from '@app/services/tools/rectangle-service';
-import { drawingToolId, sidebarToolID } from '@app/shared/enum';
+import { drawingToolId, Options, sidebarToolID } from '@app/shared/enum';
+import { BehaviorSubject } from 'rxjs';
+import { EllipseSelectorService } from './ellipse-selector-service';
 
 @Injectable({
     providedIn: 'root',
@@ -17,6 +24,7 @@ export class ToolsService {
 
     private selectedSideBarToolID: sidebarToolID;
     private currentDrawingToolID: drawingToolId;
+    toolSidenavToggle: BehaviorSubject<boolean> = new BehaviorSubject(false);
 
     constructor(
         pencilService: PencilService,
@@ -25,9 +33,30 @@ export class ToolsService {
         eraserService: EraserService,
         lineService: LineService,
         brushService: BrushService,
+        rectangleSelectionService: RectangleSelectorService,
+        ellipseSelectionService: EllipseSelectorService,
+        polygonService: PolygonService,
+        bucketService: BucketService,
+        pipetteService: PipetteService,
     ) {
         this.currentDrawingTool = pencilService;
-        this.tools = [pencilService, rectangleService, ellipseService, eraserService, lineService, brushService];
+        this.tools = [
+            pencilService,
+            rectangleService,
+            ellipseService,
+            eraserService,
+            lineService,
+            brushService,
+            rectangleSelectionService,
+            ellipseSelectionService,
+            polygonService,
+            bucketService,
+            pipetteService,
+        ];
+    }
+
+    getTool(id: drawingToolId): Tool {
+        return this.tools[id];
     }
 
     set _currentDrawingTool(newToolID: drawingToolId) {
@@ -49,5 +78,28 @@ export class ToolsService {
 
     set _selectedSideBarToolID(id: sidebarToolID) {
         this.selectedSideBarToolID = id;
+    }
+
+    toggleToolSidenav(): void {
+        this.toolSidenavToggle.next(!this.toolSidenavToggle.getValue());
+    }
+
+    openToolSidenav(): void {
+        this.toolSidenavToggle.next(true);
+    }
+
+    closeToolSidenav(): void {
+        this.toolSidenavToggle.next(false);
+    }
+
+    updateOptionValue(key: Options, option: ToolOption): void {
+        this.currentDrawingTool.options.toolOptions.set(key, option);
+    }
+
+    get currentDrawingToolOptions(): Map<Options, ToolOption> | undefined {
+        if (this.currentDrawingTool.options) {
+            return this.currentDrawingTool.options.toolOptions;
+        }
+        return undefined;
     }
 }
