@@ -1,6 +1,5 @@
-import { DateController } from '@app/controllers/date.controller';
 import { IndexController } from '@app/controllers/index.controller';
-import { DateService } from '@app/services/date.service';
+import { DatabaseService } from '@app/services/database/database.service';
 import { IndexService } from '@app/services/index.service';
 import { Container } from 'inversify';
 import { Application } from './app';
@@ -14,9 +13,7 @@ export const containerBootstrapper: () => Promise<Container> = async () => {
     container.bind(TYPES.Application).to(Application);
     container.bind(TYPES.IndexController).to(IndexController);
     container.bind(TYPES.IndexService).to(IndexService);
-
-    container.bind(TYPES.DateController).to(DateController);
-    container.bind(TYPES.DateService).to(DateService);
+    container.bind(TYPES.DatabaseService).to(DatabaseService);
 
     return container;
 };
