@@ -1,77 +1,50 @@
 import { TYPES } from '@app/types';
-import { Message } from '@common/communication/message';
+import { DrawingInfo } from '@common/communication/drawing-info';
 import { expect } from 'chai';
-import { Stubbed, testingContainer } from '../../test/test-utils';
-import { DateService } from './date.service';
+import { testingContainer } from '../../test/test-utils';
 import { IndexService } from './index.service';
 
 describe('Index service', () => {
     let indexService: IndexService;
-    let dateService: Stubbed<DateService>;
 
     beforeEach(async () => {
-        const [container, sandbox] = await testingContainer();
-        container.rebind(TYPES.DateService).toConstantValue({
-            currentTime: sandbox.stub().resolves({
-                title: 'Time',
-                body: new Date(2020, 0, 10).toString(),
-            }),
-        });
-        dateService = container.get(TYPES.DateService);
+        const [container] = await testingContainer();
         indexService = container.get<IndexService>(TYPES.IndexService);
     });
 
-    it('should return a simple message if #about is called', () => {
-        const expectedTitle = 'Basic Server About Page';
-        const expectedBody = 'Try calling helloWorld to get the time';
-        const aboutMessage = indexService.about();
-        expect(aboutMessage.title).to.equals(expectedTitle);
-        expect(aboutMessage.body).to.equals(expectedBody);
-    });
-
-    it('should return Hello World as title', (done: Mocha.Done) => {
-        indexService.helloWorld().then((result: Message) => {
-            expect(result.title).to.equals('Hello world');
-            done();
-        });
-    });
-
-    it('should have a body that starts with "Time is"', (done: Mocha.Done) => {
-        indexService.helloWorld().then((result: Message) => {
-            expect(result.body)
-                .to.be.a('string')
-                .and.satisfy((body: string) => body.startsWith('Time is'));
-            done();
-        });
-    });
-
-    it('should handle an error from DateService', (done: Mocha.Done) => {
-        dateService.currentTime.rejects(new Error('error in the service'));
-        indexService
-            .helloWorld()
-            .then((result: Message) => {
-                expect(result.title).to.equals('Error');
-                done();
-            })
-            .catch((error: unknown) => {
-                done(error);
-            });
-    });
-
-    it('should store a message', (done: Mocha.Done) => {
-        const newMessage: Message = { title: 'Hello', body: 'World' };
-        indexService.storeMessage(newMessage);
-        expect(indexService.clientMessages[0]).to.equals(newMessage);
+    it('should store a Drawing', (done: Mocha.Done) => {
+        const newDrawing: DrawingInfo = { id: 996, name: '', tags: [], metadata: '' };
+        indexService.storeDrawing(newDrawing);
+        expect(indexService.clientDrawings[0]).to.equals(newDrawing);
         done();
     });
 
-    it('should get all messages', (done: Mocha.Done) => {
-        const newMessage: Message = { title: 'Hello', body: 'World' };
-        const newMessage2: Message = { title: 'Hello', body: 'Again' };
-        indexService.clientMessages.push(newMessage);
-        indexService.clientMessages.push(newMessage2);
-        const messages = indexService.getAllMessages();
-        expect(messages).to.equals(indexService.clientMessages);
+    it('should delete a Drawing', (done: Mocha.Done) => {
+        const newDrawing: DrawingInfo = { id: 996, name: '', tags: [], metadata: '' };
+        indexService.storeDrawing(newDrawing);
+        expect(indexService.deleteDrawing(newDrawing.id).id).to.equals(newDrawing.id);
+        expect(indexService.clientDrawings.length).to.equal(0);
+        done();
+    });
+
+    it('should not delete a non-existent Drawing and return an empty drawing', (done: Mocha.Done) => {
+        const newDrawing: DrawingInfo = { id: 993, name: '', tags: [], metadata: '' };
+        const newDrawing2: DrawingInfo = { id: 994, name: '', tags: ['a'], metadata: '' };
+        indexService.storeDrawing(newDrawing);
+        indexService.storeDrawing(newDrawing2);
+        // tslint:disable-next-line: no-magic-numbers
+        indexService.deleteDrawing(995);
+        expect(indexService.clientDrawings.length).to.equals(2);
+        done();
+    });
+
+    it('should get all Drawings', (done: Mocha.Done) => {
+        const newDrawing: DrawingInfo = { id: 996, name: '', tags: [], metadata: '' };
+        const newDrawing2: DrawingInfo = { id: 996, name: '', tags: [], metadata: '' };
+        indexService.clientDrawings.push(newDrawing2);
+        indexService.clientDrawings.push(newDrawing);
+        const messages = indexService.getAllDrawings();
+        expect(messages).to.equals(indexService.clientDrawings);
         done();
     });
 });
