@@ -1,49 +1,32 @@
-import { TYPES } from '@app/types';
-import { Message } from '@common/communication/message';
-import { inject, injectable } from 'inversify';
+import { DrawingInfo } from '@common/communication/drawing-info';
+import { injectable } from 'inversify';
 import 'reflect-metadata';
-import { DateService } from './date.service';
 
 @injectable()
 export class IndexService {
-    clientMessages: Message[];
-    constructor(@inject(TYPES.DateService) private dateService: DateService) {
-        this.clientMessages = [];
+    clientDrawings: DrawingInfo[];
+    constructor() {
+        this.clientDrawings = [];
     }
 
-    about(): Message {
-        return {
-            title: 'Basic Server About Page',
-            body: 'Try calling helloWorld to get the time',
-        };
+    storeDrawing(drawing: DrawingInfo): void {
+        this.clientDrawings.push(drawing);
     }
 
-    async helloWorld(): Promise<Message> {
-        return this.dateService
-            .currentTime()
-            .then((timeMessage: Message) => {
-                return {
-                    title: 'Hello world',
-                    body: 'Time is ' + timeMessage.body,
-                };
-            })
-            .catch((error: unknown) => {
-                console.error('There was an error!!!', error);
-
-                return {
-                    title: 'Error',
-                    body: error as string,
-                };
-            });
+    getAllDrawings(): DrawingInfo[] {
+        return this.clientDrawings;
     }
 
-    // TODO : ceci est à titre d'exemple. À enlever pour la remise
-    storeMessage(message: Message): void {
-        console.log(message);
-        this.clientMessages.push(message);
-    }
-
-    getAllMessages(): Message[] {
-        return this.clientMessages;
+    deleteDrawing(drawingId: number): DrawingInfo {
+        let deletedDrawing: DrawingInfo;
+        deletedDrawing = { id: -1, name: 'drawingNotFound', tags: [''], metadata: '' };
+        for (const clientDrawing of this.clientDrawings) {
+            if (clientDrawing.id === drawingId) {
+                const index = this.clientDrawings.indexOf(clientDrawing);
+                this.clientDrawings.splice(index, 1);
+                deletedDrawing = clientDrawing;
+            }
+        }
+        return deletedDrawing;
     }
 }
