@@ -2,7 +2,6 @@ export const TYPES = {
     Server: Symbol('Server'),
     Application: Symbol('Application'),
     IndexController: Symbol('IndexController'),
-    DateController: Symbol('DateController'),
     IndexService: Symbol('IndexService'),
-    DateService: Symbol('DateService'),
+    DatabaseService: Symbol('DatabaseService'),
 };
